test(e2e): move Brave setup branches to helpers
